requirements.txt
pyvisa
pyvisa-sim
paho-mqtt
fastapi
uvicorn
pyyaml
numpy
pandas
scikit-learn
openai
langchain
tiktoken
